refactor: extract locale logic to src/utils/server/locale.ts
